Replace tokenizer with simple line parser
I realized that the ini format is just based on lines which do not
require an advanced parser to interpret what a simple line parser
could do.

Surely this new version is a little bit more strict about the ini file
format but since it seems to correlate with Microsofts own specification
I am absolutely fine with that.