Add ability to write ini formated data to string
From time to time it might be useful to write an RSettings object back
into file after it was changed by the application.

The original order and comments are not preserved at the moment. Which
means that one should not invest much time in formating and describing
the ini file if he plans to write it back from the application.

Value pairs for which only the default value is retrieved, will also not
be written into the string.