Add error handling for ill formated ini data
This patch adds the Result value type, which can be used to notify the
user about errors in the submitted ini formated data.

The Result object can be casted to bool whereas true means that the
given ini data could be parsed properly. In an error case the reasen
contains the line number and a human readable message, why the data
could not be parsed.